refactor: 마시기 목표 설정 UI 개선
