+++ android/app/src/main/AndroidManifest.xml
@@ -60,7 +60,7 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".ui.settinggoal.SettingGoalActivity"
+            android:name=".ui.settingtargetamount.SettingTargetAmountActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
