+++ android/app/src/main/java/com/mulkkam/ui/setting/SettingFragment.kt
@@ -12,8 +12,8 @@ import com.mulkkam.ui.main.Refreshable
 import com.mulkkam.ui.setting.adapter.SettingAdapter
 import com.mulkkam.ui.setting.adapter.SettingItem
 import com.mulkkam.ui.setting.model.SettingType
-import com.mulkkam.ui.settinggoal.SettingGoalActivity
 import com.mulkkam.ui.settingnickname.SettingNicknameActivity
+import com.mulkkam.ui.settingtargetamount.SettingTargetAmountActivity
 import com.mulkkam.ui.settingwater.SettingWaterActivity
 
 class SettingFragment :
@@ -55,7 +55,7 @@ class SettingFragment :
             }
 
             SettingType.Normal.MyCup -> startActivity(SettingWaterActivity.newIntent(requireContext()))
-            SettingType.Normal.Goal -> startActivity(SettingGoalActivity.newIntent(requireContext()))
+            SettingType.Normal.Goal -> startActivity(SettingTargetAmountActivity.newIntent(requireContext()))
             SettingType.Normal.Notification -> navigateToNotificationSetting()
         }
     }
